Aux data gaps reporting endpoint, error/retry scenes order link
